+++ src/utils/error_handler.c
@@ -11,7 +11,7 @@ void	log_error(char *str_error, int status_error)
 {
 	ft_putstr_fd(MS_ERROR, STDERR_FILENO);
 	ft_putstr_fd(str_error, STDERR_FILENO);
-	g_var.last_cmd_status = status_error;
+	g_var.current_status = status_error;
 }
 
 /**