+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2022-01-25 - Edu
+### Added
+- Added **current_status** **last_status** to global to to handle error expansions
+
 ## 2022-01-24 - Jorge
 
 ### Fixed